refactor(testability): rename function to get testability
Previously, getting testability was `window.angular2.getTestability`
This was because the plan was to export the API to the window as
angular2. However, the decision was changed to make this just `angular`
in https://github.com/angular/angular/commit/2795c010dc0fcf11e896a0a9c74f967e61bf7841

To decouple testability from the rest of the Angular API, just make it
one function, `window.getAngularTestability`.