docs(developer.md): add PUB_CACHE variable to setup instructions
If PUB_CACHE is not set the pub dependencies are downloaded to a default
folder during the build because `gulp build` also runs `pub get`.

Running the tests locally PUB_CACHE is set to a subdirectory of the SDK
 when the `env_dart.sh` script is sourced.

Therefore `build/pubbuild.dart` fails as it can't find the dependencies.

Closes #4900